requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0

# Text processing
nltk>=3.6.0
emoji>=1.6.0
contractions>=0.1.72

# ML enhancements
imbalanced-learn>=0.8.0  # for SMOTE

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Web framework (for the API)
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2  # Template engine

# Development and testing
pytest>=6.2.0
black>=21.6b0
flake8>=3.9.0
mypy>=0.910
httpx  # HTTP client for testing

# Logging and monitoring
python-json-logger>=2.0.0
prometheus-fastapi-instrumentator  # Monitoring tool

# Type checking
typing-extensions>=3.10.0

# Environment management
python-dotenv  # Load environment variables
